@@ -5,11 +5,9 @@ author: Flutter Team <flutter-dev@googlegroups.com>
 dependencies:
   flutter:
     sdk: flutter
-  cupertino_icons: ^0.1.2
-  shared_preferences: ^0.5.2
+  shared_preferences: ^2.0.0-nullsafety.1
 
 dev_dependencies:
-  test: ^1.5.2
   flutter_driver:
     sdk: flutter
   in_app_purchase:
@@ -21,11 +19,11 @@ dev_dependencies:
     path: ../
   integration_test:
     path: ../../integration_test
-  pedantic: ^1.8.0
+  pedantic: ^1.10.0
 
 flutter:
   uses-material-design: true
 
 environment:
-  sdk: ">=2.3.0 <3.0.0"
+  sdk: ">=2.12.0-259.9.beta <3.0.0"
   flutter: ">=1.9.1+hotfix.2"
