@@ -1,30 +1,26 @@
 name: in_app_purchase
 description: A Flutter plugin for in-app purchases. Exposes APIs for making in-app purchases through the App Store and Google Play.
 homepage: https://github.com/flutter/plugins/tree/master/packages/in_app_purchase
-version: 0.3.5+2
+version: 0.4.0
 
 dependencies:
-  async: ^2.0.8
-  collection: ^1.14.11
   flutter:
     sdk: flutter
-  json_annotation: ^3.0.0
-  meta: ^1.1.6
+  json_annotation: ^4.0.0
+  meta: ^1.3.0
+  collection: ^1.15.0
 
 dev_dependencies:
-  build_runner: ^1.0.0
-  json_serializable: ^3.2.0
+  build_runner: ^1.11.1
+  json_serializable: ^4.0.0
   flutter_test:
     sdk: flutter
   flutter_driver:
     sdk: flutter
-  in_app_purchase_example:
-    path: example/
-  test: ^1.5.2
-  shared_preferences: ^0.5.2
+  test: ^1.16.0
   integration_test:
     path: ../integration_test
-  pedantic: ^1.8.0
+  pedantic: ^1.10.0
 
 flutter:
   plugin:
@@ -36,5 +32,5 @@ flutter:
         pluginClass: InAppPurchasePlugin
 
 environment:
-  sdk: ">=2.3.0 <3.0.0"
+  sdk: ">=2.12.0-259.9.beta <3.0.0"
   flutter: ">=1.12.13+hotfix.5"
