@@ -12,40 +12,50 @@ part 'enum_converters.g.dart';
 ///
 /// Use these in `@JsonSerializable()` classes by annotating them with
 /// `@BillingResponseConverter()`.
-class BillingResponseConverter implements JsonConverter<BillingResponse, int> {
+class BillingResponseConverter implements JsonConverter<BillingResponse, int?> {
   /// Default const constructor.
   const BillingResponseConverter();
 
   @override
-  BillingResponse fromJson(int json) => _$enumDecode<BillingResponse>(
-      _$BillingResponseEnumMap.cast<BillingResponse, dynamic>(), json);
+  BillingResponse fromJson(int? json) {
+    if (json == null) {
+      return BillingResponse.error;
+    }
+    return _$enumDecode<BillingResponse, dynamic>(
+        _$BillingResponseEnumMap.cast<BillingResponse, dynamic>(), json);
+  }
 
   @override
-  int toJson(BillingResponse object) => _$BillingResponseEnumMap[object];
+  int toJson(BillingResponse object) => _$BillingResponseEnumMap[object]!;
 }
 
 /// Serializer for [SkuType].
 ///
 /// Use these in `@JsonSerializable()` classes by annotating them with
 /// `@SkuTypeConverter()`.
-class SkuTypeConverter implements JsonConverter<SkuType, String> {
+class SkuTypeConverter implements JsonConverter<SkuType, String?> {
   /// Default const constructor.
   const SkuTypeConverter();
 
   @override
-  SkuType fromJson(String json) =>
-      _$enumDecode<SkuType>(_$SkuTypeEnumMap.cast<SkuType, dynamic>(), json);
+  SkuType fromJson(String? json) {
+    if (json == null) {
+      return SkuType.inapp;
+    }
+    return _$enumDecode<SkuType, dynamic>(
+        _$SkuTypeEnumMap.cast<SkuType, dynamic>(), json);
+  }
 
   @override
-  String toJson(SkuType object) => _$SkuTypeEnumMap[object];
+  String toJson(SkuType object) => _$SkuTypeEnumMap[object]!;
 }
 
 // Define a class so we generate serializer helper methods for the enums
 @JsonSerializable()
 class _SerializedEnums {
-  BillingResponse response;
-  SkuType type;
-  PurchaseStateWrapper purchaseState;
+  late BillingResponse response;
+  late SkuType type;
+  late PurchaseStateWrapper purchaseState;
 }
 
 /// Serializer for [PurchaseStateWrapper].
@@ -53,18 +63,23 @@ class _SerializedEnums {
 /// Use these in `@JsonSerializable()` classes by annotating them with
 /// `@PurchaseStateConverter()`.
 class PurchaseStateConverter
-    implements JsonConverter<PurchaseStateWrapper, int> {
+    implements JsonConverter<PurchaseStateWrapper, int?> {
   /// Default const constructor.
   const PurchaseStateConverter();
 
   @override
-  PurchaseStateWrapper fromJson(int json) => _$enumDecode<PurchaseStateWrapper>(
-      _$PurchaseStateWrapperEnumMap.cast<PurchaseStateWrapper, dynamic>(),
-      json);
+  PurchaseStateWrapper fromJson(int? json) {
+    if (json == null) {
+      return PurchaseStateWrapper.unspecified_state;
+    }
+    return _$enumDecode<PurchaseStateWrapper, dynamic>(
+        _$PurchaseStateWrapperEnumMap.cast<PurchaseStateWrapper, dynamic>(),
+        json);
+  }
 
   @override
   int toJson(PurchaseStateWrapper object) =>
-      _$PurchaseStateWrapperEnumMap[object];
+      _$PurchaseStateWrapperEnumMap[object]!;
 
   /// Converts the purchase state stored in `object` to a [PurchaseStatus].
   ///
@@ -78,7 +93,5 @@ class PurchaseStateConverter
       case PurchaseStateWrapper.unspecified_state:
         return PurchaseStatus.error;
     }
-
-    throw ArgumentError('$object isn\'t mapped to PurchaseStatus');
   }
 }
