@@ -21,9 +21,9 @@ class AppStoreConnection implements InAppPurchaseConnection {
   /// Returns the singleton instance of the [AppStoreConnection] that should be
   /// used across the app.
   static AppStoreConnection get instance => _getOrCreateInstance();
-  static AppStoreConnection _instance;
-  static SKPaymentQueueWrapper _skPaymentQueueWrapper;
-  static _TransactionObserver _observer;
+  static AppStoreConnection? _instance;
+  static late SKPaymentQueueWrapper _skPaymentQueueWrapper;
+  static late _TransactionObserver _observer;
 
   /// Creates an [AppStoreConnection] object.
   ///
@@ -41,55 +41,61 @@ class AppStoreConnection implements InAppPurchaseConnection {
 
   static AppStoreConnection _getOrCreateInstance() {
     if (_instance != null) {
-      return _instance;
+      return _instance!;
     }
 
     _instance = AppStoreConnection();
     _skPaymentQueueWrapper = SKPaymentQueueWrapper();
     _observer = _TransactionObserver(StreamController.broadcast());
     _skPaymentQueueWrapper.setTransactionObserver(observer);
-    return _instance;
+    return _instance!;
   }
 
   @override
   Future<bool> isAvailable() => SKPaymentQueueWrapper.canMakePayments();
 
   @override
-  Future<bool> buyNonConsumable({@required PurchaseParam purchaseParam}) async {
+  Future<bool> buyNonConsumable({required PurchaseParam purchaseParam}) async {
     await _skPaymentQueueWrapper.addPayment(SKPaymentWrapper(
         productIdentifier: purchaseParam.productDetails.id,
         quantity: 1,
         applicationUsername: purchaseParam.applicationUserName,
-        simulatesAskToBuyInSandbox: purchaseParam.sandboxTesting,
+        simulatesAskToBuyInSandbox: purchaseParam.simulatesAskToBuyInSandbox ||
+            // ignore: deprecated_member_use_from_same_package
+            purchaseParam.sandboxTesting,
         requestData: null));
     return true; // There's no error feedback from iOS here to return.
   }
 
   @override
   Future<bool> buyConsumable(
-      {@required PurchaseParam purchaseParam, bool autoConsume = true}) {
+      {required PurchaseParam purchaseParam, bool autoConsume = true}) {
     assert(autoConsume == true, 'On iOS, we should always auto consume');
     return buyNonConsumable(purchaseParam: purchaseParam);
   }
 
   @override
   Future<BillingResultWrapper> completePurchase(PurchaseDetails purchase,
-      {String developerPayload}) async {
+      {String? developerPayload}) async {
+    if (purchase.skPaymentTransaction == null) {
+      throw ArgumentError(
+          'completePurchase unsuccessful. The `purchase.skPaymentTransaction` is not valid');
+    }
     await _skPaymentQueueWrapper
-        .finishTransaction(purchase.skPaymentTransaction);
+        .finishTransaction(purchase.skPaymentTransaction!);
     return BillingResultWrapper(responseCode: BillingResponse.ok);
   }
 
   @override
   Future<BillingResultWrapper> consumePurchase(PurchaseDetails purchase,
-      {String developerPayload}) {
+      {String? developerPayload}) {
     throw UnsupportedError('consume purchase is not available on Android');
   }
 
   @override
   Future<QueryPurchaseDetailsResponse> queryPastPurchases(
-      {String applicationUserName}) async {
-    IAPError error;
+      {String? applicationUserName}) async {
+    IAPError? error;
     List<PurchaseDetails> pastPurchases = [];
 
     try {
@@ -98,7 +104,6 @@ class AppStoreConnection implements InAppPurchaseConnection {
           await _observer.getRestoredTransactions(
               queue: _skPaymentQueueWrapper,
               applicationUserName: applicationUserName);
-      _observer.cleanUpRestoredTransactions();
       pastPurchases =
           restoredTransactions.map((SKPaymentTransactionWrapper transaction) {
         assert(transaction.transactionState ==
@@ -110,16 +115,17 @@ class AppStoreConnection implements InAppPurchaseConnection {
               ? IAPError(
                   source: IAPSource.AppStore,
                   code: kPurchaseErrorCode,
-                  message: transaction.error.domain,
-                  details: transaction.error.userInfo,
+                  message: transaction.error?.domain ?? '',
+                  details: transaction.error?.userInfo,
                 )
               : null;
       }).toList();
+      _observer.cleanUpRestoredTransactions();
     } on PlatformException catch (e) {
       error = IAPError(
           source: IAPSource.AppStore,
           code: e.code,
-          message: e.message,
+          message: e.message ?? '',
           details: e.details);
     } on SKError catch (e) {
       error = IAPError(
@@ -133,9 +139,12 @@ class AppStoreConnection implements InAppPurchaseConnection {
   }
 
   @override
-  Future<PurchaseVerificationData> refreshPurchaseVerificationData() async {
+  Future<PurchaseVerificationData?> refreshPurchaseVerificationData() async {
     await SKRequestMaker().startRefreshReceiptRequest();
-    String receipt = await SKReceiptManager.retrieveReceiptData();
+    final String? receipt = await SKReceiptManager.retrieveReceiptData();
+    if (receipt == null) {
+      return null;
+    }
     return PurchaseVerificationData(
         localVerificationData: receipt,
         serverVerificationData: receipt,
@@ -152,7 +161,7 @@ class AppStoreConnection implements InAppPurchaseConnection {
       Set<String> identifiers) async {
     final SKRequestMaker requestMaker = SKRequestMaker();
     SkProductResponseWrapper response;
-    PlatformException exception;
+    PlatformException? exception;
     try {
       response = await requestMaker.startProductRequest(identifiers.toList());
     } on PlatformException catch (e) {
@@ -167,7 +176,7 @@ class AppStoreConnection implements InAppPurchaseConnection {
               ProductDetails.fromSKProduct(productWrapper))
           .toList();
     }
-    List<String> invalidIdentifiers = response.invalidProductIdentifiers ?? [];
+    List<String> invalidIdentifiers = response.invalidProductIdentifiers;
     if (productDetails.isEmpty) {
       invalidIdentifiers = identifiers.toList();
     }
@@ -179,7 +188,7 @@ class AppStoreConnection implements InAppPurchaseConnection {
           : IAPError(
               source: IAPSource.AppStore,
               code: exception.code,
-              message: exception.message,
+              message: exception.message ?? '',
               details: exception.details),
     );
     return productDetailsResponse;
@@ -189,27 +198,27 @@ class AppStoreConnection implements InAppPurchaseConnection {
 class _TransactionObserver implements SKTransactionObserverWrapper {
   final StreamController<List<PurchaseDetails>> purchaseUpdatedController;
 
-  Completer<List<SKPaymentTransactionWrapper>> _restoreCompleter;
-  List<SKPaymentTransactionWrapper> _restoredTransactions;
-  String _receiptData;
+  Completer<List<SKPaymentTransactionWrapper>>? _restoreCompleter;
+  List<SKPaymentTransactionWrapper> _restoredTransactions =
+      <SKPaymentTransactionWrapper>[];
+  late String _receiptData;
 
   _TransactionObserver(this.purchaseUpdatedController);
 
   Future<List<SKPaymentTransactionWrapper>> getRestoredTransactions(
-      {@required SKPaymentQueueWrapper queue, String applicationUserName}) {
-    assert(queue != null);
+      {required SKPaymentQueueWrapper queue, String? applicationUserName}) {
     _restoreCompleter = Completer();
     queue.restoreTransactions(applicationUserName: applicationUserName);
-    return _restoreCompleter.future;
+    return _restoreCompleter!.future;
   }
 
   void cleanUpRestoredTransactions() {
-    _restoredTransactions = null;
+    _restoredTransactions.clear();
     _restoreCompleter = null;
   }
 
   void updatedTransactions(
-      {List<SKPaymentTransactionWrapper> transactions}) async {
+      {required List<SKPaymentTransactionWrapper> transactions}) async {
     if (_restoreCompleter != null) {
       if (_restoredTransactions == null) {
         _restoredTransactions = [];
@@ -233,19 +242,20 @@ class _TransactionObserver implements SKTransactionObserverWrapper {
     }).toList());
   }
 
-  void removedTransactions({List<SKPaymentTransactionWrapper> transactions}) {}
+  void removedTransactions(
+      {required List<SKPaymentTransactionWrapper> transactions}) {}
 
   /// Triggered when there is an error while restoring transactions.
-  void restoreCompletedTransactionsFailed({SKError error}) {
-    _restoreCompleter.completeError(error);
+  void restoreCompletedTransactionsFailed({required SKError error}) {
+    _restoreCompleter!.completeError(error);
   }
 
   void paymentQueueRestoreCompletedTransactionsFinished() {
-    _restoreCompleter.complete(_restoredTransactions ?? []);
+    _restoreCompleter!.complete(_restoredTransactions);
   }
 
   bool shouldAddStorePayment(
-      {SKPaymentWrapper payment, SKProductWrapper product}) {
+      {required SKPaymentWrapper payment, required SKProductWrapper product}) {
     // In this unified API, we always return true to keep it consistent with the behavior on Google Play.
     return true;
   }
@@ -254,7 +264,7 @@ class _TransactionObserver implements SKTransactionObserverWrapper {
     try {
       _receiptData = await SKReceiptManager.retrieveReceiptData();
     } catch (e) {
-      _receiptData = null;
+      _receiptData = '';
     }
     return _receiptData;
   }
