@@ -8,6 +8,7 @@ import 'package:flutter/services.dart';
 import 'package:flutter/widgets.dart';
 import 'package:in_app_purchase/src/in_app_purchase/purchase_details.dart';
 import '../../billing_client_wrappers.dart';
+import '../../in_app_purchase.dart';
 import 'in_app_purchase_connection.dart';
 import 'product_details.dart';
 
@@ -28,26 +29,27 @@ class GooglePlayConnection
       billingClient.enablePendingPurchases();
     }
     _readyFuture = _connect();
-    WidgetsBinding.instance.addObserver(this);
+    WidgetsBinding.instance!.addObserver(this);
     _purchaseUpdatedController = StreamController.broadcast();
     ;
   }
 
   /// Returns the singleton instance of the [GooglePlayConnection].
   static GooglePlayConnection get instance => _getOrCreateInstance();
-  static GooglePlayConnection _instance;
+  static GooglePlayConnection? _instance;
 
   Stream<List<PurchaseDetails>> get purchaseUpdatedStream =>
       _purchaseUpdatedController.stream;
-  static StreamController<List<PurchaseDetails>> _purchaseUpdatedController;
+  static late StreamController<List<PurchaseDetails>>
+      _purchaseUpdatedController;
 
   /// The [BillingClient] that's abstracted by [GooglePlayConnection].
   ///
   /// This field should not be used out of test code.
   @visibleForTesting
-  final BillingClient billingClient;
+  late final BillingClient billingClient;
 
-  Future<void> _readyFuture;
+  late Future<void> _readyFuture;
   static Set<String> _productIdsToConsume = Set<String>();
 
   @override
@@ -57,7 +59,7 @@ class GooglePlayConnection
   }
 
   @override
-  Future<bool> buyNonConsumable({@required PurchaseParam purchaseParam}) async {
+  Future<bool> buyNonConsumable({required PurchaseParam purchaseParam}) async {
     BillingResultWrapper billingResultWrapper =
         await billingClient.launchBillingFlow(
             sku: purchaseParam.productDetails.id,
@@ -67,7 +69,7 @@ class GooglePlayConnection
 
   @override
   Future<bool> buyConsumable(
-      {@required PurchaseParam purchaseParam, bool autoConsume = true}) {
+      {required PurchaseParam purchaseParam, bool autoConsume = true}) {
     if (autoConsume) {
       _productIdsToConsume.add(purchaseParam.productDetails.id);
     }
@@ -76,9 +78,13 @@ class GooglePlayConnection
 
   @override
   Future<BillingResultWrapper> completePurchase(PurchaseDetails purchase,
-      {String developerPayload}) async {
-    if (purchase.billingClientPurchase.isAcknowledged) {
+      {String? developerPayload}) async {
+    if (purchase.billingClientPurchase!.isAcknowledged) {
       return BillingResultWrapper(responseCode: BillingResponse.ok);
+    }
+    if (purchase.verificationData == null) {
+      throw ArgumentError(
+          'completePurchase unsuccessful. The `purchase.verificationData` is not valid');
     }
     return await billingClient.acknowledgePurchase(
         purchase.verificationData.serverVerificationData,
@@ -87,7 +93,11 @@ class GooglePlayConnection
 
   @override
   Future<BillingResultWrapper> consumePurchase(PurchaseDetails purchase,
-      {String developerPayload}) {
+      {String? developerPayload}) {
+    if (purchase.verificationData == null) {
+      throw ArgumentError(
+          'consumePurchase unsuccessful. The `purchase.verificationData` is not valid');
+    }
     return billingClient.consumeAsync(
         purchase.verificationData.serverVerificationData,
         developerPayload: developerPayload);
@@ -95,9 +105,9 @@ class GooglePlayConnection
 
   @override
   Future<QueryPurchaseDetailsResponse> queryPastPurchases(
-      {String applicationUserName}) async {
+      {String? applicationUserName}) async {
     List<PurchasesResultWrapper> responses;
-    PlatformException exception;
+    PlatformException? exception;
     try {
       responses = await Future.wait([
         billingClient.queryPurchases(SkuType.inapp),
@@ -133,7 +143,7 @@ class GooglePlayConnection
         .toSet();
 
     String errorMessage =
-        errorCodeSet.isNotEmpty ? errorCodeSet.join(', ') : null;
+        errorCodeSet.isNotEmpty ? errorCodeSet.join(', ') : '';
 
     List<PurchaseDetails> pastPurchases =
         responses.expand((PurchasesResultWrapper response) {
@@ -142,14 +152,14 @@ class GooglePlayConnection
       return PurchaseDetails.fromPurchase(purchaseWrapper);
     }).toList();
 
-    IAPError error;
+    IAPError? error;
     if (exception != null) {
       error = IAPError(
           source: IAPSource.GooglePlay,
           code: exception.code,
-          message: exception.message,
+          message: exception.message ?? '',
           details: exception.details);
-    } else if (errorMessage != null) {
+    } else if (errorMessage.isNotEmpty) {
       error = IAPError(
           source: IAPSource.GooglePlay,
           code: kRestoredPurchaseErrorCode,
@@ -175,11 +185,11 @@ class GooglePlayConnection
 
   static GooglePlayConnection _getOrCreateInstance() {
     if (_instance != null) {
-      return _instance;
+      return _instance!;
     }
 
     _instance = GooglePlayConnection._();
-    return _instance;
+    return _instance!;
   }
 
   Future<void> _connect() =>
@@ -193,7 +203,7 @@ class GooglePlayConnection
   Future<ProductDetailsResponse> queryProductDetails(
       Set<String> identifiers) async {
     List<SkuDetailsResponseWrapper> responses;
-    PlatformException exception;
+    PlatformException? exception;
     try {
       responses = await Future.wait([
         billingClient.querySkuDetails(
@@ -235,13 +245,13 @@ class GooglePlayConnection
             : IAPError(
                 source: IAPSource.GooglePlay,
                 code: exception.code,
-                message: exception.message,
+                message: exception.message ?? '',
                 details: exception.details));
   }
 
   static Future<List<PurchaseDetails>> _getPurchaseDetailsFromResult(
       PurchasesResultWrapper resultWrapper) async {
-    IAPError error;
+    IAPError? error;
     if (resultWrapper.responseCode != BillingResponse.ok) {
       error = IAPError(
         source: IAPSource.GooglePlay,
@@ -260,10 +270,13 @@ class GooglePlayConnection
     } else {
       return [
         PurchaseDetails(
-            purchaseID: null,
-            productID: null,
+            purchaseID: '',
+            productID: '',
             transactionDate: null,
-            verificationData: null)
+            verificationData: PurchaseVerificationData(
+                localVerificationData: '',
+                serverVerificationData: '',
+                source: IAPSource.GooglePlay))
           ..status = PurchaseStatus.error
           ..error = error
       ];
