@@ -2,7 +2,6 @@
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 
-import 'package:flutter/foundation.dart';
 import 'package:in_app_purchase/store_kit_wrappers.dart';
 import 'package:in_app_purchase/billing_client_wrappers.dart';
 import 'in_app_purchase_connection.dart';
@@ -14,10 +13,10 @@ import 'in_app_purchase_connection.dart';
 class ProductDetails {
   /// Creates a new product details object with the provided details.
   ProductDetails(
-      {@required this.id,
-      @required this.title,
-      @required this.description,
-      @required this.price,
+      {required this.id,
+      required this.title,
+      required this.description,
+      required this.price,
       this.skProduct,
       this.skuDetail});
 
@@ -36,13 +35,13 @@ class ProductDetails {
 
   /// Points back to the `StoreKits`'s [SKProductWrapper] object that generated this [ProductDetails] object.
   ///
-  /// This is null on Android.
-  final SKProductWrapper skProduct;
+  /// This is `null` on Android.
+  final SKProductWrapper? skProduct;
 
   /// Points back to the `BillingClient1`'s [SkuDetailsWrapper] object that generated this [ProductDetails] object.
   ///
-  /// This is null on iOS.
-  final SkuDetailsWrapper skuDetail;
+  /// This is `null` on iOS.
+  final SkuDetailsWrapper? skuDetail;
 
   /// Generate a [ProductDetails] object based on an iOS [SKProductWrapper] object.
   ProductDetails.fromSKProduct(SKProductWrapper product)
@@ -69,7 +68,7 @@ class ProductDetails {
 class ProductDetailsResponse {
   /// Creates a new [ProductDetailsResponse] with the provided response details.
   ProductDetailsResponse(
-      {@required this.productDetails, @required this.notFoundIDs, this.error});
+      {required this.productDetails, required this.notFoundIDs, this.error});
 
   /// Each [ProductDetails] uniquely matches one valid identifier in [identifiers] of [InAppPurchaseConnection.queryProductDetails].
   final List<ProductDetails> productDetails;
@@ -82,7 +81,9 @@ class ProductDetailsResponse {
 
   /// A caught platform exception thrown while querying the purchases.
   ///
+  /// The value is `null` if there is no error.
+  ///
   /// It's possible for this to be null but for there still to be notFoundIds in cases where the request itself was a success but the
   /// requested IDs could not be found.
-  final IAPError error;
+  final IAPError? error;
 }
