@@ -15,6 +15,7 @@ import 'package:in_app_purchase/src/in_app_purchase/app_store_connection.dart';
 import 'package:in_app_purchase/src/in_app_purchase/in_app_purchase_connection.dart';
 import 'package:in_app_purchase/src/in_app_purchase/product_details.dart';
 import 'package:in_app_purchase/store_kit_wrappers.dart';
+import '../billing_client_wrappers/purchase_wrapper_test.dart';
 import '../store_kit_wrappers/sk_test_stub_objects.dart';
 
 void main() {
@@ -61,10 +62,11 @@ void main() {
           .queryProductDetails(<String>['123', '456', '789'].toSet());
       expect(response.productDetails, []);
       expect(response.notFoundIDs, ['123', '456', '789']);
-      expect(response.error.source, IAPSource.AppStore);
-      expect(response.error.code, 'error_code');
-      expect(response.error.message, 'error_message');
-      expect(response.error.details, {'info': 'error_info'});
+      expect(response.error, isNotNull);
+      expect(response.error!.source, IAPSource.AppStore);
+      expect(response.error!.code, 'error_code');
+      expect(response.error!.message, 'error_message');
+      expect(response.error!.details, {'info': 'error_info'});
     });
   });
 
@@ -81,6 +83,8 @@ void main() {
           fakeIOSPlatform.transactions.first.transactionIdentifier);
       expect(response.pastPurchases.last.purchaseID,
           fakeIOSPlatform.transactions.last.transactionIdentifier);
+      expect(response.pastPurchases, isNotEmpty);
+      expect(response.pastPurchases.first.verificationData, isNotNull);
       expect(
           response.pastPurchases.first.verificationData.localVerificationData,
           'dummy base64data');
@@ -97,7 +101,7 @@ void main() {
       Stream<List<PurchaseDetails>> stream =
           AppStoreConnection.instance.purchaseUpdatedStream;
 
-      StreamSubscription subscription;
+      late StreamSubscription subscription;
       subscription = stream.listen((purchaseDetailsList) {
         if (purchaseDetailsList.first.status == PurchaseStatus.purchased) {
           completer.complete(purchaseDetailsList);
@@ -130,9 +134,10 @@ void main() {
       QueryPurchaseDetailsResponse response =
           await AppStoreConnection.instance.queryPastPurchases();
       expect(response.pastPurchases, isEmpty);
-      expect(response.error.source, IAPSource.AppStore);
-      expect(response.error.message, 'error_test');
-      expect(response.error.details, {'message': 'errorMessage'});
+      expect(response.error, isNotNull);
+      expect(response.error!.source, IAPSource.AppStore);
+      expect(response.error!.message, 'error_test');
+      expect(response.error!.details, {'message': 'errorMessage'});
     });
 
     test('receipt error should populate null to verificationData.data',
@@ -142,18 +147,19 @@ void main() {
           await AppStoreConnection.instance.queryPastPurchases();
       expect(
           response.pastPurchases.first.verificationData.localVerificationData,
-          null);
+          isEmpty);
       expect(
           response.pastPurchases.first.verificationData.serverVerificationData,
-          null);
+          isEmpty);
     });
   });
 
   group('refresh receipt data', () {
     test('should refresh receipt data', () async {
-      PurchaseVerificationData receiptData =
+      PurchaseVerificationData? receiptData =
           await AppStoreConnection.instance.refreshPurchaseVerificationData();
-      expect(receiptData.source, IAPSource.AppStore);
+      expect(receiptData, isNotNull);
+      expect(receiptData!.source, IAPSource.AppStore);
       expect(receiptData.localVerificationData, 'refreshed receipt data');
       expect(receiptData.serverVerificationData, 'refreshed receipt data');
     });
@@ -168,7 +174,7 @@ void main() {
       Stream<List<PurchaseDetails>> stream =
           AppStoreConnection.instance.purchaseUpdatedStream;
 
-      StreamSubscription subscription;
+      late StreamSubscription subscription;
       subscription = stream.listen((purchaseDetailsList) {
         details.addAll(purchaseDetailsList);
         if (purchaseDetailsList.first.status == PurchaseStatus.purchased) {
@@ -195,7 +201,7 @@ void main() {
       Stream<List<PurchaseDetails>> stream =
           AppStoreConnection.instance.purchaseUpdatedStream;
 
-      StreamSubscription subscription;
+      late StreamSubscription subscription;
       subscription = stream.listen((purchaseDetailsList) {
         details.addAll(purchaseDetailsList);
         if (purchaseDetailsList.first.status == PurchaseStatus.purchased) {
@@ -228,16 +234,16 @@ void main() {
       fakeIOSPlatform.testTransactionFail = true;
       List<PurchaseDetails> details = [];
       Completer completer = Completer();
-      IAPError error;
+      late IAPError error;
 
       Stream<List<PurchaseDetails>> stream =
           AppStoreConnection.instance.purchaseUpdatedStream;
-      StreamSubscription subscription;
+      late StreamSubscription subscription;
       subscription = stream.listen((purchaseDetailsList) {
         details.addAll(purchaseDetailsList);
         purchaseDetailsList.forEach((purchaseDetails) {
           if (purchaseDetails.status == PurchaseStatus.error) {
-            error = purchaseDetails.error;
+            error = purchaseDetails.error!;
             completer.complete(error);
             subscription.cancel();
           }
@@ -263,7 +269,7 @@ void main() {
       Completer completer = Completer();
       Stream<List<PurchaseDetails>> stream =
           AppStoreConnection.instance.purchaseUpdatedStream;
-      StreamSubscription subscription;
+      late StreamSubscription subscription;
       subscription = stream.listen((purchaseDetailsList) {
         details.addAll(purchaseDetailsList);
         purchaseDetailsList.forEach((purchaseDetails) {
@@ -288,7 +294,9 @@ void main() {
 
   group('consume purchase', () {
     test('should throw when calling consume purchase on iOS', () async {
-      expect(() => AppStoreConnection.instance.consumePurchase(null),
+      expect(
+          () => AppStoreConnection.instance
+              .consumePurchase(PurchaseDetails.fromPurchase(dummyPurchase)),
           throwsUnsupportedError);
     });
   });
@@ -300,16 +308,16 @@ class FakeIOSPlatform {
   }
 
   // pre-configured store informations
-  String receiptData;
-  Set<String> validProductIDs;
-  Map<String, SKProductWrapper> validProducts;
-  List<SKPaymentTransactionWrapper> transactions;
-  List<SKPaymentTransactionWrapper> finishedTransactions;
-  bool testRestoredTransactionsNull;
-  bool testTransactionFail;
-  PlatformException queryProductException;
-  PlatformException restoreException;
-  SKError testRestoredError;
+  String? receiptData;
+  late Set<String> validProductIDs;
+  late Map<String, SKProductWrapper> validProducts;
+  late List<SKPaymentTransactionWrapper> transactions;
+  late List<SKPaymentTransactionWrapper> finishedTransactions;
+  late bool testRestoredTransactionsNull;
+  late bool testTransactionFail;
+  PlatformException? queryProductException;
+  PlatformException? restoreException;
+  SKError? testRestoredError;
 
   void reset() {
     transactions = [];
@@ -317,7 +325,8 @@ class FakeIOSPlatform {
     validProductIDs = ['123', '456'].toSet();
     validProducts = Map();
     for (String validID in validProductIDs) {
-      Map productWrapperMap = buildProductMap(dummyProductWrapper);
+      Map<String, dynamic> productWrapperMap =
+          buildProductMap(dummyProductWrapper);
       productWrapperMap['productIdentifier'] = validID;
       validProducts[validID] = SKProductWrapper.fromJson(productWrapperMap);
     }
@@ -350,7 +359,7 @@ class FakeIOSPlatform {
 
   SKPaymentTransactionWrapper createPendingTransaction(String id) {
     return SKPaymentTransactionWrapper(
-        transactionIdentifier: null,
+        transactionIdentifier: '',
         payment: SKPaymentWrapper(productIdentifier: id),
         transactionState: SKPaymentTransactionStateWrapper.purchasing,
         transactionTimeStamp: 123123.121,
@@ -371,7 +380,7 @@ class FakeIOSPlatform {
 
   SKPaymentTransactionWrapper createFailedTransaction(String productId) {
     return SKPaymentTransactionWrapper(
-        transactionIdentifier: null,
+        transactionIdentifier: '',
         payment: SKPaymentWrapper(productIdentifier: productId),
         transactionState: SKPaymentTransactionStateWrapper.failed,
         transactionTimeStamp: 123123.121,
@@ -388,7 +397,7 @@ class FakeIOSPlatform {
         return Future<bool>.value(true);
       case '-[InAppPurchasePlugin startProductRequest:result:]':
         if (queryProductException != null) {
-          throw queryProductException;
+          throw queryProductException!;
         }
         List<String> productIDS =
             List.castFrom<dynamic, String>(call.arguments);
@@ -399,7 +408,7 @@ class FakeIOSPlatform {
           if (!validProductIDs.contains(productID)) {
             invalidFound.add(productID);
           } else {
-            products.add(validProducts[productID]);
+            products.add(validProducts[productID]!);
           }
         }
         SkProductResponseWrapper response = SkProductResponseWrapper(
@@ -408,11 +417,11 @@ class FakeIOSPlatform {
             buildProductResponseMap(response));
       case '-[InAppPurchasePlugin restoreTransactions:result:]':
         if (restoreException != null) {
-          throw restoreException;
+          throw restoreException!;
         }
         if (testRestoredError != null) {
           AppStoreConnection.observer
-              .restoreCompletedTransactionsFailed(error: testRestoredError);
+              .restoreCompletedTransactionsFailed(error: testRestoredError!);
           return Future<void>.sync(() {});
         }
         if (!testRestoredTransactionsNull) {
@@ -428,7 +437,6 @@ class FakeIOSPlatform {
         } else {
           throw PlatformException(code: 'no_receipt_data');
         }
-        break;
       case '-[InAppPurchasePlugin refreshReceipt:result:]':
         receiptData = 'refreshed receipt data';
         return Future<void>.sync(() {});
@@ -445,7 +453,8 @@ class FakeIOSPlatform {
               .updatedTransactions(transactions: [transaction_failed]);
         } else {
           SKPaymentTransactionWrapper transaction_finished =
-              createPurchasedTransaction(id, transaction.transactionIdentifier);
+              createPurchasedTransaction(
+                  id, transaction.transactionIdentifier ?? '');
           AppStoreConnection.observer
               .updatedTransactions(transactions: [transaction_finished]);
         }
