@@ -24,7 +24,7 @@ void main() {
   TestWidgetsFlutterBinding.ensureInitialized();
 
   final StubInAppPurchasePlatform stubPlatform = StubInAppPurchasePlatform();
-  GooglePlayConnection connection;
+  late GooglePlayConnection connection;
   const String startConnectionCall =
       'BillingClient#startConnection(BillingClientStateListener)';
   const String endConnectionCall = 'BillingClient#endConnection()';
@@ -149,10 +149,11 @@ void main() {
           await connection.queryProductDetails(<String>['invalid'].toSet());
       expect(response.notFoundIDs, ['invalid']);
       expect(response.productDetails, isEmpty);
-      expect(response.error.source, IAPSource.GooglePlay);
-      expect(response.error.code, 'error_code');
-      expect(response.error.message, 'error_message');
-      expect(response.error.details, {'info': 'error_info'});
+      expect(response.error, isNotNull);
+      expect(response.error!.source, IAPSource.GooglePlay);
+      expect(response.error!.code, 'error_code');
+      expect(response.error!.message, 'error_message');
+      expect(response.error!.details, {'info': 'error_info'});
     });
   });
 
@@ -172,8 +173,10 @@ void main() {
       final QueryPurchaseDetailsResponse response =
           await connection.queryPastPurchases();
       expect(response.pastPurchases, isEmpty);
-      expect(response.error.message, BillingResponse.developerError.toString());
-      expect(response.error.source, IAPSource.GooglePlay);
+      expect(response.error, isNotNull);
+      expect(
+          response.error!.message, BillingResponse.developerError.toString());
+      expect(response.error!.source, IAPSource.GooglePlay);
     });
 
     test('returns SkuDetailsResponseWrapper', () async {
@@ -221,9 +224,10 @@ void main() {
       final QueryPurchaseDetailsResponse response =
           await connection.queryPastPurchases();
       expect(response.pastPurchases, isEmpty);
-      expect(response.error.code, 'error_code');
-      expect(response.error.message, 'error_message');
-      expect(response.error.details, {'info': 'error_info'});
+      expect(response.error, isNotNull);
+      expect(response.error!.code, 'error_code');
+      expect(response.error!.message, 'error_message');
+      expect(response.error!.details, {'info': 'error_info'});
     });
   });
 
@@ -277,7 +281,7 @@ void main() {
       PurchaseDetails purchaseDetails;
       Stream purchaseStream =
           GooglePlayConnection.instance.purchaseUpdatedStream;
-      StreamSubscription subscription;
+      late StreamSubscription subscription;
       subscription = purchaseStream.listen((_) {
         purchaseDetails = _.first;
         completer.complete(purchaseDetails);
@@ -320,7 +324,7 @@ void main() {
       PurchaseDetails purchaseDetails;
       Stream purchaseStream =
           GooglePlayConnection.instance.purchaseUpdatedStream;
-      StreamSubscription subscription;
+      late StreamSubscription subscription;
       subscription = purchaseStream.listen((_) {
         purchaseDetails = _.first;
         completer.complete(purchaseDetails);
@@ -334,9 +338,9 @@ void main() {
       PurchaseDetails result = await completer.future;
 
       expect(result.error, isNotNull);
-      expect(result.error.source, IAPSource.GooglePlay);
+      expect(result.error!.source, IAPSource.GooglePlay);
       expect(result.status, PurchaseStatus.error);
-      expect(result.purchaseID, isNull);
+      expect(result.purchaseID, isEmpty);
     });
 
     test('buy consumable with auto consume, serializes and deserializes data',
@@ -392,7 +396,7 @@ void main() {
       PurchaseDetails purchaseDetails;
       Stream purchaseStream =
           GooglePlayConnection.instance.purchaseUpdatedStream;
-      StreamSubscription subscription;
+      late StreamSubscription subscription;
       subscription = purchaseStream.listen((_) {
         purchaseDetails = _.first;
         completer.complete(purchaseDetails);
@@ -407,7 +411,8 @@ void main() {
       // Verify that the result has succeeded
       PurchaseDetails result = await completer.future;
       expect(launchResult, isTrue);
-      expect(result.billingClientPurchase.purchaseToken,
+      expect(result.billingClientPurchase, isNotNull);
+      expect(result.billingClientPurchase!.purchaseToken,
           await consumeCompleter.future);
       expect(result.status, PurchaseStatus.purchased);
       expect(result.error, isNull);
@@ -501,7 +506,7 @@ void main() {
       PurchaseDetails purchaseDetails;
       Stream purchaseStream =
           GooglePlayConnection.instance.purchaseUpdatedStream;
-      StreamSubscription subscription;
+      late StreamSubscription subscription;
       subscription = purchaseStream.listen((_) {
         purchaseDetails = _.first;
         completer.complete(purchaseDetails);
@@ -515,11 +520,12 @@ void main() {
 
       // Verify that the result has an error for the failed consumption
       PurchaseDetails result = await completer.future;
-      expect(result.billingClientPurchase.purchaseToken,
+      expect(result.billingClientPurchase, isNotNull);
+      expect(result.billingClientPurchase!.purchaseToken,
           await consumeCompleter.future);
       expect(result.status, PurchaseStatus.error);
       expect(result.error, isNotNull);
-      expect(result.error.code, kConsumptionFailedErrorCode);
+      expect(result.error!.code, kConsumptionFailedErrorCode);
     });
 
     test(
@@ -574,7 +580,7 @@ void main() {
 
       Stream purchaseStream =
           GooglePlayConnection.instance.purchaseUpdatedStream;
-      StreamSubscription subscription;
+      late StreamSubscription subscription;
       subscription = purchaseStream.listen((_) {
         consumeCompleter.complete(null);
         subscription.cancel();
@@ -629,10 +635,6 @@ void main() {
             await GooglePlayConnection.instance.completePurchase(
                 purchaseDetails,
                 developerPayload: 'dummy payload');
-        print('pending ${billingResultWrapper.responseCode}');
-        print('expectedBillingResult ${expectedBillingResult.responseCode}');
-        print('pending ${billingResultWrapper.debugMessage}');
-        print('expectedBillingResult ${expectedBillingResult.debugMessage}');
         expect(billingResultWrapper, equals(expectedBillingResult));
         completer.complete(billingResultWrapper);
       }
