@@ -8,22 +8,22 @@ part of 'sku_details_wrapper.dart';
 
 SkuDetailsWrapper _$SkuDetailsWrapperFromJson(Map json) {
   return SkuDetailsWrapper(
-    description: json['description'] as String,
-    freeTrialPeriod: json['freeTrialPeriod'] as String,
-    introductoryPrice: json['introductoryPrice'] as String,
-    introductoryPriceMicros: json['introductoryPriceMicros'] as String,
-    introductoryPriceCycles: json['introductoryPriceCycles'] as String,
-    introductoryPricePeriod: json['introductoryPricePeriod'] as String,
-    price: json['price'] as String,
-    priceAmountMicros: json['priceAmountMicros'] as int,
-    priceCurrencyCode: json['priceCurrencyCode'] as String,
-    sku: json['sku'] as String,
-    subscriptionPeriod: json['subscriptionPeriod'] as String,
-    title: json['title'] as String,
-    type: const SkuTypeConverter().fromJson(json['type'] as String),
-    isRewarded: json['isRewarded'] as bool,
-    originalPrice: json['originalPrice'] as String,
-    originalPriceAmountMicros: json['originalPriceAmountMicros'] as int,
+    description: json['description'] as String? ?? '',
+    freeTrialPeriod: json['freeTrialPeriod'] as String? ?? '',
+    introductoryPrice: json['introductoryPrice'] as String? ?? '',
+    introductoryPriceMicros: json['introductoryPriceMicros'] as String? ?? '',
+    introductoryPriceCycles: json['introductoryPriceCycles'] as String? ?? '',
+    introductoryPricePeriod: json['introductoryPricePeriod'] as String? ?? '',
+    price: json['price'] as String? ?? '',
+    priceAmountMicros: json['priceAmountMicros'] as int? ?? 0,
+    priceCurrencyCode: json['priceCurrencyCode'] as String? ?? '',
+    sku: json['sku'] as String? ?? '',
+    subscriptionPeriod: json['subscriptionPeriod'] as String? ?? '',
+    title: json['title'] as String? ?? '',
+    type: const SkuTypeConverter().fromJson(json['type'] as String?),
+    isRewarded: json['isRewarded'] as bool? ?? false,
+    originalPrice: json['originalPrice'] as String? ?? '',
+    originalPriceAmountMicros: json['originalPriceAmountMicros'] as int? ?? 0,
   );
 }
 
@@ -49,10 +49,15 @@ Map<String, dynamic> _$SkuDetailsWrapperToJson(SkuDetailsWrapper instance) =>
 
 SkuDetailsResponseWrapper _$SkuDetailsResponseWrapperFromJson(Map json) {
   return SkuDetailsResponseWrapper(
-    billingResult: BillingResultWrapper.fromJson(json['billingResult'] as Map),
-    skuDetailsList: (json['skuDetailsList'] as List)
-        .map((e) => SkuDetailsWrapper.fromJson(e as Map))
-        .toList(),
+    billingResult:
+        BillingResultWrapper.fromJson((json['billingResult'] as Map?)?.map(
+      (k, e) => MapEntry(k as String, e),
+    )),
+    skuDetailsList: (json['skuDetailsList'] as List<dynamic>?)
+            ?.map((e) =>
+                SkuDetailsWrapper.fromJson(Map<String, dynamic>.from(e as Map)))
+            .toList() ??
+        [],
   );
 }
 
@@ -66,8 +71,8 @@ Map<String, dynamic> _$SkuDetailsResponseWrapperToJson(
 BillingResultWrapper _$BillingResultWrapperFromJson(Map json) {
   return BillingResultWrapper(
     responseCode:
-        const BillingResponseConverter().fromJson(json['responseCode'] as int),
-    debugMessage: json['debugMessage'] as String,
+        const BillingResponseConverter().fromJson(json['responseCode'] as int?),
+    debugMessage: json['debugMessage'] as String?,
   );
 }
 
