@@ -8,12 +8,16 @@ part of 'sk_product_wrapper.dart';
 
 SkProductResponseWrapper _$SkProductResponseWrapperFromJson(Map json) {
   return SkProductResponseWrapper(
-    products: (json['products'] as List)
-        .map((e) => SKProductWrapper.fromJson(e as Map))
-        .toList(),
-    invalidProductIdentifiers: (json['invalidProductIdentifiers'] as List)
-        .map((e) => e as String)
-        .toList(),
+    products: (json['products'] as List<dynamic>?)
+            ?.map((e) =>
+                SKProductWrapper.fromJson(Map<String, dynamic>.from(e as Map)))
+            .toList() ??
+        [],
+    invalidProductIdentifiers:
+        (json['invalidProductIdentifiers'] as List<dynamic>?)
+                ?.map((e) => e as String)
+                .toList() ??
+            [],
   );
 }
 
@@ -27,8 +31,9 @@ Map<String, dynamic> _$SkProductResponseWrapperToJson(
 SKProductSubscriptionPeriodWrapper _$SKProductSubscriptionPeriodWrapperFromJson(
     Map json) {
   return SKProductSubscriptionPeriodWrapper(
-    numberOfUnits: json['numberOfUnits'] as int,
-    unit: _$enumDecodeNullable(_$SKSubscriptionPeriodUnitEnumMap, json['unit']),
+    numberOfUnits: json['numberOfUnits'] as int? ?? 0,
+    unit: const SKSubscriptionPeriodUnitConverter()
+        .fromJson(json['unit'] as int?),
   );
 }
 
@@ -36,61 +41,23 @@ Map<String, dynamic> _$SKProductSubscriptionPeriodWrapperToJson(
         SKProductSubscriptionPeriodWrapper instance) =>
     <String, dynamic>{
       'numberOfUnits': instance.numberOfUnits,
-      'unit': _$SKSubscriptionPeriodUnitEnumMap[instance.unit],
+      'unit': const SKSubscriptionPeriodUnitConverter().toJson(instance.unit),
     };
-
-T _$enumDecode<T>(
-  Map<T, dynamic> enumValues,
-  dynamic source, {
-  T unknownValue,
-}) {
-  if (source == null) {
-    throw ArgumentError('A value must be provided. Supported values: '
-        '${enumValues.values.join(', ')}');
-  }
-
-  final value = enumValues.entries
-      .singleWhere((e) => e.value == source, orElse: () => null)
-      ?.key;
-
-  if (value == null && unknownValue == null) {
-    throw ArgumentError('`$source` is not one of the supported values: '
-        '${enumValues.values.join(', ')}');
-  }
-  return value ?? unknownValue;
-}
-
-T _$enumDecodeNullable<T>(
-  Map<T, dynamic> enumValues,
-  dynamic source, {
-  T unknownValue,
-}) {
-  if (source == null) {
-    return null;
-  }
-  return _$enumDecode<T>(enumValues, source, unknownValue: unknownValue);
-}
-
-const _$SKSubscriptionPeriodUnitEnumMap = {
-  SKSubscriptionPeriodUnit.day: 0,
-  SKSubscriptionPeriodUnit.week: 1,
-  SKSubscriptionPeriodUnit.month: 2,
-  SKSubscriptionPeriodUnit.year: 3,
-};
 
 SKProductDiscountWrapper _$SKProductDiscountWrapperFromJson(Map json) {
   return SKProductDiscountWrapper(
-    price: json['price'] as String,
-    priceLocale: json['priceLocale'] == null
-        ? null
-        : SKPriceLocaleWrapper.fromJson(json['priceLocale'] as Map),
-    numberOfPeriods: json['numberOfPeriods'] as int,
-    paymentMode: _$enumDecodeNullable(
-        _$SKProductDiscountPaymentModeEnumMap, json['paymentMode']),
-    subscriptionPeriod: json['subscriptionPeriod'] == null
-        ? null
-        : SKProductSubscriptionPeriodWrapper.fromJson(
-            json['subscriptionPeriod'] as Map),
+    price: json['price'] as String? ?? '',
+    priceLocale:
+        SKPriceLocaleWrapper.fromJson((json['priceLocale'] as Map?)?.map(
+      (k, e) => MapEntry(k as String, e),
+    )),
+    numberOfPeriods: json['numberOfPeriods'] as int? ?? 0,
+    paymentMode: const SKProductDiscountPaymentModeConverter()
+        .fromJson(json['paymentMode'] as int?),
+    subscriptionPeriod: SKProductSubscriptionPeriodWrapper.fromJson(
+        (json['subscriptionPeriod'] as Map?)?.map(
+      (k, e) => MapEntry(k as String, e),
+    )),
   );
 }
 
@@ -100,34 +67,32 @@ Map<String, dynamic> _$SKProductDiscountWrapperToJson(
       'price': instance.price,
       'priceLocale': instance.priceLocale,
       'numberOfPeriods': instance.numberOfPeriods,
-      'paymentMode':
-          _$SKProductDiscountPaymentModeEnumMap[instance.paymentMode],
+      'paymentMode': const SKProductDiscountPaymentModeConverter()
+          .toJson(instance.paymentMode),
       'subscriptionPeriod': instance.subscriptionPeriod,
     };
 
-const _$SKProductDiscountPaymentModeEnumMap = {
-  SKProductDiscountPaymentMode.payAsYouGo: 0,
-  SKProductDiscountPaymentMode.payUpFront: 1,
-  SKProductDiscountPaymentMode.freeTrail: 2,
-};
-
 SKProductWrapper _$SKProductWrapperFromJson(Map json) {
   return SKProductWrapper(
-    productIdentifier: json['productIdentifier'] as String,
-    localizedTitle: json['localizedTitle'] as String,
-    localizedDescription: json['localizedDescription'] as String,
-    priceLocale: json['priceLocale'] == null
-        ? null
-        : SKPriceLocaleWrapper.fromJson(json['priceLocale'] as Map),
-    subscriptionGroupIdentifier: json['subscriptionGroupIdentifier'] as String,
-    price: json['price'] as String,
+    productIdentifier: json['productIdentifier'] as String? ?? '',
+    localizedTitle: json['localizedTitle'] as String? ?? '',
+    localizedDescription: json['localizedDescription'] as String? ?? '',
+    priceLocale:
+        SKPriceLocaleWrapper.fromJson((json['priceLocale'] as Map?)?.map(
+      (k, e) => MapEntry(k as String, e),
+    )),
+    subscriptionGroupIdentifier: json['subscriptionGroupIdentifier'] as String?,
+    price: json['price'] as String? ?? '',
     subscriptionPeriod: json['subscriptionPeriod'] == null
         ? null
         : SKProductSubscriptionPeriodWrapper.fromJson(
-            json['subscriptionPeriod'] as Map),
+            (json['subscriptionPeriod'] as Map?)?.map(
+            (k, e) => MapEntry(k as String, e),
+          )),
     introductoryPrice: json['introductoryPrice'] == null
         ? null
-        : SKProductDiscountWrapper.fromJson(json['introductoryPrice'] as Map),
+        : SKProductDiscountWrapper.fromJson(
+            Map<String, dynamic>.from(json['introductoryPrice'] as Map)),
   );
 }
 
@@ -145,8 +110,8 @@ Map<String, dynamic> _$SKProductWrapperToJson(SKProductWrapper instance) =>
 
 SKPriceLocaleWrapper _$SKPriceLocaleWrapperFromJson(Map json) {
   return SKPriceLocaleWrapper(
-    currencySymbol: json['currencySymbol'] as String,
-    currencyCode: json['currencyCode'] as String,
+    currencySymbol: json['currencySymbol'] as String? ?? '',
+    currencyCode: json['currencyCode'] as String? ?? '',
   );
 }
 
