@@ -24,7 +24,7 @@ class SKRequestMaker {
   /// A [PlatformException] is thrown if the platform code making the request fails.
   Future<SkProductResponseWrapper> startProductRequest(
       List<String> productIdentifiers) async {
-    final Map<String, dynamic> productResponseMap =
+    final Map<String, dynamic>? productResponseMap =
         await channel.invokeMapMethod<String, dynamic>(
       '-[InAppPurchasePlugin startProductRequest:result:]',
       productIdentifiers,
@@ -47,7 +47,8 @@ class SKRequestMaker {
   /// * isExpired: whether the receipt is expired.
   /// * isRevoked: whether the receipt has been revoked.
   /// * isVolumePurchase: whether the receipt is a Volume Purchase Plan receipt.
-  Future<void> startRefreshReceiptRequest({Map receiptProperties}) {
+  Future<void> startRefreshReceiptRequest(
+      {Map<String, dynamic>? receiptProperties}) {
     return channel.invokeMethod<void>(
       '-[InAppPurchasePlugin refreshReceipt:result:]',
       receiptProperties,
