@@ -9,19 +9,19 @@ typedef void AdditionalSteps(dynamic args);
 
 class StubInAppPurchasePlatform {
   Map<String, dynamic> _expectedCalls = <String, dynamic>{};
-  Map<String, AdditionalSteps> _additionalSteps = <String, AdditionalSteps>{};
+  Map<String, AdditionalSteps?> _additionalSteps = <String, AdditionalSteps?>{};
   void addResponse(
-      {String name,
+      {required String name,
       dynamic value,
-      AdditionalSteps additionalStepBeforeReturn}) {
+      AdditionalSteps? additionalStepBeforeReturn}) {
     _additionalSteps[name] = additionalStepBeforeReturn;
     _expectedCalls[name] = value;
   }
 
   List<MethodCall> _previousCalls = <MethodCall>[];
   List<MethodCall> get previousCalls => _previousCalls;
-  MethodCall previousCallMatching(String name) => _previousCalls
-      .firstWhere((MethodCall call) => call.method == name, orElse: () => null);
+  MethodCall previousCallMatching(String name) =>
+      _previousCalls.firstWhere((MethodCall call) => call.method == name);
   int countPreviousCalls(String name) =>
       _previousCalls.where((MethodCall call) => call.method == name).length;
 
@@ -35,7 +35,7 @@ class StubInAppPurchasePlatform {
     _previousCalls.add(call);
     if (_expectedCalls.containsKey(call.method)) {
       if (_additionalSteps[call.method] != null) {
-        _additionalSteps[call.method](call.arguments);
+        _additionalSteps[call.method]!(call.arguments);
       }
       return Future<dynamic>.sync(() => _expectedCalls[call.method]);
     } else {
